// src/app/modules/reports/email-bill/models/ebill-search-result.model.ts
export class EbillSearchResult{
    acknowledgeCount: number;
    billGeneratedCount: number;
    billingLevel: string;
    billingLevelCode: string;
    customerName: string;
    emailSentCount: number;
    msaCode: string;
    rateCardCode: string;
    sendingFailedCount: number;
    sfxCode: string;
}